Previously it had very subtle accessors, like the -> operator to get the result model (or the call () operator in Python). It also allowed you to use the output object directly as an iterable container.
Now there is an explicit "test_run" member that gives access to the test run model, which reads better like "output.test_run.start" instead of "output.start", and is consistent with "test_step" for accessing the test step model.

Also now there is a "raw" member that gives access to the raw output artifact protos for iteration.

PiperOrigin-RevId: 490753868